Move List into New Todo List Component
