+++ src/App.jsx
@@ -2,19 +2,9 @@ import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
+import TodoList from './TodoList'
 
-const todoList = [{
-  "id": "1",
-  "title": "Complete Assignment"
-},
-{
-  "id": "2",
-  "title": "Read textbook materials",
-  },
-{
-  "id": "3",
-  "title": "Stretch "
-}]
+
 
 
 function App() {
@@ -23,13 +13,7 @@ function App() {
   return (
     <>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map(function (item) {
-          return (
-            <li key={item.id}>{item.title}</li>
-          );
-        })}
-      </ul>
+      <TodoList />
     </>
   )
 }
